# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Offix API",
    "description": "API REST del backend de Offix, red social de oficios (UTN FRSF).\n\n### Cómo autenticarse\n1. `POST /api/v1/auth/registro` para crear un usuario.\n2. `POST /api/v1/auth/login` para obtener un `access_token` (JWT).\n3. Enviarlo en cada request protegida como header `Authorization: Bearer <access_token>` (en Swagger UI: botón **Authorize**, arriba a la derecha).\n\n### Para el equipo de QA (Hoppscotch)\nEl spec completo también está disponible en `/openapi.json` y como archivo estático en `docs/openapi.json` del repo — se puede importar directo en Hoppscotch como colección.\n\n### Estado del proyecto\nSprint 1 — flujo de reseña (HU-01) con validación de enlace simulada, sin envío real de mail/WhatsApp. Ver `docs/plan-sprint-1.md` para el detalle de alcance y decisiones.",
    "contact": {
      "name": "Equipo Offix - UTN FRSF"
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Entorno local (iniciar_backend.bat)"
    }
  ],
  "paths": {
    "/api/v1/auth/registro": {
      "post": {
        "tags": [
          "Autenticación"
        ],
        "summary": "Registrar un nuevo usuario",
        "description": "RF1/RF4 — Registro de nuevos usuarios Oferentes.\n\nCrea el Usuario (rol Oferente por defecto). Para poder generar/aprobar\nreseñas o gestionar un perfil profesional, después hay que crear el\nperfil con `POST /api/v1/oferentes` usando el token devuelto por `/login`.",
        "operationId": "registrar_usuario_api_v1_auth_registro_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UsuarioCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsuarioOut"
                }
              }
            }
          },
          "409": {
            "description": "El email ya está registrado"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "Autenticación"
        ],
        "summary": "Iniciar sesión",
        "description": "RF4 — Inicio de sesión.\n\nDevuelve un JWT (`access_token`). Enviarlo en cada request protegida como\nheader `Authorization: Bearer <token>`.",
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UsuarioLogin"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Token"
                }
              }
            }
          },
          "401": {
            "description": "Email o contraseña incorrectos"
          },
          "403": {
            "description": "La cuenta está suspendida o bloqueada"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "tags": [
          "Autenticación"
        ],
        "summary": "Cerrar sesión",
        "description": "HU-03 T03 — El JWT es stateless y no hay tabla de sesiones/blacklist en el\nDER acordado con la PM, así que no se revoca el token en el servidor. Este\nendpoint exige un token válido (confirma que había sesión activa); el cierre\nde sesión real lo hace el cliente descartando el token guardado.",
        "operationId": "logout_api_v1_auth_logout_post",
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "description": "Token faltante, inválido o expirado"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/oferentes": {
      "get": {
        "tags": [
          "Oferentes"
        ],
        "summary": "Buscar profesionales (público)",
        "description": "RF7/RF8 — Búsqueda y consulta pública de perfiles profesionales. No requiere login.",
        "operationId": "buscar_oferentes_api_v1_oferentes_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Palabra clave: nombre o apellido",
              "title": "Q"
            },
            "description": "Palabra clave: nombre o apellido"
          },
          {
            "name": "categoria_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filtrar por categoría/oficio",
              "title": "Categoria Id"
            },
            "description": "Filtrar por categoría/oficio"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OferenteOut"
                  },
                  "title": "Response Buscar Oferentes Api V1 Oferentes Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Oferentes"
        ],
        "summary": "Crear el perfil profesional del usuario autenticado",
        "description": "RF2/HU-02 — Creación del perfil profesional del oferente. El\n`id_oferente` resultante es siempre igual al `id_usuario` del usuario\nautenticado (comparten clave primaria).",
        "operationId": "crear_perfil_oferente_api_v1_oferentes_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OferenteCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OferenteOut"
                }
              }
            }
          },
          "401": {
            "description": "Falta token o es inválido"
          },
          "403": {
            "description": "El usuario autenticado no tiene rol Oferente"
          },
          "409": {
            "description": "El usuario ya tiene perfil, o el DNI/CUIT ya está registrado"
          },
          "422": {
            "description": "Campo obligatorio faltante o DNI/CUIT/teléfono con formato inválido"
          }
        }
      }
    },
...
    "/api/v1/admin/usuarios": {
      "get": {
        "tags": [
          "Administración"
        ],
        "summary": "Listar todos los usuarios",
        "description": "RF14 — Gestión de usuarios.",
        "operationId": "listar_usuarios_api_v1_admin_usuarios_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/UsuarioOut"
                  },
                  "type": "array",
                  "title": "Response Listar Usuarios Api V1 Admin Usuarios Get"
                }
              }
            }
          },
          "401": {
            "description": "Falta token o es inválido"
          },
          "403": {
            "description": "Requiere rol Administrador"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/admin/usuarios/{usuario_id}/estado": {
      "patch": {
        "tags": [
          "Administración"
        ],
        "summary": "Suspender/bloquear/reactivar un usuario",
        "description": "RF15 — Bloquear o suspender usuarios. Un usuario Suspendida/Bloqueada no puede hacer login.",
        "operationId": "cambiar_estado_usuario_api_v1_admin_usuarios__usuario_id__estado_patch",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "usuario_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Usuario Id"
            }
          },
          {
            "name": "estado",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/EstadoCuentaEnum"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsuarioOut"
                }
              }
            }
          },
          "401": {
            "description": "Falta token o es inválido"
          },
          "403": {
            "description": "Requiere rol Administrador"
          },
          "404": {
            "description": "Usuario no encontrado"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/oferentes/{oferente_id}/verificacion": {
      "patch": {
        "tags": [
          "Administración"
        ],
        "summary": "Marcar manualmente el estado de verificación de matrícula",
        "description": "HU-02 T07 — Simula la integración con el ente validador de matrículas\n(electricista, gasista, plomero, técnico en aire acondicionado): para\nSprint 1 el estado se marca manualmente en vez de consultar un servicio\nexterno real.",
        "operationId": "actualizar_estado_verificacion_api_v1_admin_oferentes__oferente_id__verificacion_patch",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "oferente_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Oferente Id"
            }
          },
          {
            "name": "estado_verificacion",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/EstadoVerificacionEnum"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OferenteOut"
                }
              }
            }
          },
          "401": {
            "description": "Falta token o es inválido"
          },
          "403": {
            "description": "Requiere rol Administrador"
          },
          "404": {
            "description": "Oferente no encontrado"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/alertas": {
      "get": {
        "tags": [
          "Administración"
        ],
        "summary": "Listar alertas generadas por rechazos reiterados de reseñas",
        "description": "RF17 — Revisión de alertas por rechazos reiterados de reseñas.",
        "operationId": "listar_alertas_api_v1_admin_alertas_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Falta token o es inválido"
          },
          "403": {
            "description": "Requiere rol Administrador"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/admin/resenas/rechazadas": {
      "get": {
        "tags": [
          "Administración"
        ],
        "summary": "Listar reseñas rechazadas por sus oferentes",
        "description": "RF17 — El administrador puede revisar y revertir rechazos incorrectos.",
        "operationId": "listar_resenas_rechazadas_api_v1_admin_resenas_rechazadas_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ResenaAdminOut"
                  },
                  "type": "array",
                  "title": "Response Listar Resenas Rechazadas Api V1 Admin Resenas Rechazadas Get"
                }
              }
            }
          },
          "401": {
            "description": "Falta token o es inválido"
          },
          "403": {
            "description": "Requiere rol Administrador"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/admin/resenas/{resena_id}/publicar": {
      "patch": {
        "tags": [
          "Administración"
        ],
        "summary": "Publicar una reseña que el oferente había rechazado",
        "description": "RF17 — El Administrador, como instancia final, publica una reseña rechazada por el Oferente.",
        "operationId": "publicar_resena_rechazada_api_v1_admin_resenas__resena_id__publicar_patch",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "resena_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Resena Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResenaAdminOut"
                }
              }
            }
          },
          "401": {
            "description": "Falta token o es inválido"
          },
          "403": {
            "description": "Requiere rol Administrador"
          },
          "404": {
            "description": "Reseña no encontrada"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EstadoCuentaEnum": {
        "type": "string",
        "enum": [
          "Activa",
          "Suspendida",
          "Bloqueada"
        ],
        "title": "EstadoCuentaEnum"
      },
      "EstadoVerificacionEnum": {
        "type": "string",
        "enum": [
          "Pendiente",
          "Verificado",
          "Rechazado"
        ],
        "title": "EstadoVerificacionEnum"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "OferenteCreate": {
        "properties": {
          "nombre": {
            "type": "string",
            "title": "Nombre"
          },
          "apellido": {
            "type": "string",
            "title": "Apellido"
          },
          "dni_cuit": {
            "type": "string",
            "title": "Dni Cuit"
          },
          "telefono": {
            "type": "string",
            "title": "Telefono"
          },
          "categoria_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Categoria Id"
          },
          "numero_matricula": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Numero Matricula"
          },
          "latitud": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latitud"
          },
          "longitud": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Longitud"
          },
          "radio_cobertura_km": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Radio Cobertura Km"
          },
          "hora_inicio_atencion": {
            "anyOf": [
              {
                "type": "string",
                "format": "time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hora Inicio Atencion"
          },
          "hora_fin_atencion": {
            "anyOf": [
              {
                "type": "string",
                "format": "time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hora Fin Atencion"
          },
          "disponible_emergencia": {
            "type": "boolean",
            "title": "Disponible Emergencia",
            "default": false
          },
          "descripcion": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Descripcion"
          }
        },
        "type": "object",
        "required": [
          "nombre",
          "apellido",
          "dni_cuit",
          "telefono"
        ],
        "title": "OferenteCreate",
        "example": {
          "apellido": "Pérez",
          "categoria_id": 1,
          "descripcion": "Electricista matriculado, más de 10 años de experiencia.",
          "disponible_emergencia": true,
          "dni_cuit": "20-30111222-3",
          "nombre": "Juan",
          "numero_matricula": "EL-12345",
          "telefono": "+54 3564 400111"
        }
      },
      "OferenteOut": {
        "properties": {
          "nombre": {
            "type": "string",
            "title": "Nombre"
          },
          "apellido": {
            "type": "string",
            "title": "Apellido"
          },
          "dni_cuit": {
            "type": "string",
            "title": "Dni Cuit"
          },
          "telefono": {
            "type": "string",
            "title": "Telefono"
          },
          "categoria_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Categoria Id"
          },
          "numero_matricula": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Numero Matricula"
          },
          "latitud": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latitud"
          },
          "longitud": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Longitud"
          },
          "radio_cobertura_km": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Radio Cobertura Km"
          },
          "hora_inicio_atencion": {
            "anyOf": [
              {
                "type": "string",
                "format": "time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hora Inicio Atencion"
          },
          "hora_fin_atencion": {
            "anyOf": [
              {
                "type": "string",
                "format": "time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hora Fin Atencion"
          },
          "disponible_emergencia": {
            "type": "boolean",
            "title": "Disponible Emergencia",
            "default": false
          },
          "descripcion": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Descripcion"
          },
          "id_oferente": {
            "type": "integer",
            "title": "Id Oferente"
          },
          "estado_verificacion": {
            "type": "string",
            "title": "Estado Verificacion"
          },
          "promedio_calificacion": {
            "type": "string",
            "title": "Promedio Calificacion"
          },
          "cantidad_rechazos_acumulados": {
            "type": "integer",
            "title": "Cantidad Rechazos Acumulados"
          }
        },
        "type": "object",
        "required": [
          "nombre",
          "apellido",
          "dni_cuit",
          "telefono",
          "id_oferente",
          "estado_verificacion",
          "promedio_calificacion",
          "cantidad_rechazos_acumulados"
        ],
        "title": "OferenteOut"
      },
...
      "ResenaAdminOut": {
        "properties": {
          "id_resena": {
            "type": "integer",
            "title": "Id Resena"
          },
          "oferente_id": {
            "type": "integer",
            "title": "Oferente Id"
          },
          "solicitud_id": {
            "type": "integer",
            "title": "Solicitud Id"
          },
          "nombre_cliente": {
            "type": "string",
            "title": "Nombre Cliente"
          },
          "calificaciones_comentarios": {
            "type": "object",
            "title": "Calificaciones Comentarios"
          },
          "estado": {
            "type": "string",
            "title": "Estado",
            "description": "Pendiente_Aprobacion | Aprobada | Rechazada"
          },
          "fecha_creacion": {
            "type": "string",
            "format": "date-time",
            "title": "Fecha Creacion"
          },
          "contacto_cliente_ingresado": {
            "type": "string",
            "title": "Contacto Cliente Ingresado"
          }
        },
        "type": "object",
        "required": [
          "id_resena",
          "oferente_id",
          "solicitud_id",
          "nombre_cliente",
          "calificaciones_comentarios",
          "estado",
          "fecha_creacion",
          "contacto_cliente_ingresado"
        ],
        "title": "ResenaAdminOut"
      },
...
      "Token": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "bearer"
          }
        },
        "type": "object",
        "required": [
          "access_token"
        ],
        "title": "Token"
      },
      "UsuarioCreate": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "UsuarioCreate",
        "example": {
          "email": "juan.perez@offix.test",
          "password": "Offix2026!"
        }
      },
      "UsuarioLogin": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "UsuarioLogin",
        "example": {
          "email": "juan.perez@offix.test",
          "password": "Offix2026!"
        }
      },
      "UsuarioOut": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "id_usuario": {
            "type": "integer",
            "title": "Id Usuario"
          },
          "rol": {
            "$ref": "#/components/schemas/RolUsuario"
          },
          "estado_cuenta": {
            "type": "string",
            "title": "Estado Cuenta"
          },
          "fecha_registro": {
            "type": "string",
            "format": "date-time",
            "title": "Fecha Registro"
          }
        },
        "type": "object",
        "required": [
          "email",
          "id_usuario",
          "rol",
          "estado_cuenta",
          "fecha_registro"
        ],
        "title": "UsuarioOut"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  },
...
}